refactor: depend on pre-registration-core only for identity migration
