+++ pre-registration-booking-service/pom.xml
@@ -14,7 +14,6 @@
 		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
 		<io.micrometer.prometheus.version>1.4.2</io.micrometer.prometheus.version>
 		<pre.registration.core.version>1.4.0-SNAPSHOT</pre.registration.core.version>
-		<pre.registration.application.service.version>1.4.0-SNAPSHOT</pre.registration.application.service.version>
 		<kernel.core.version>1.3.1-rc.1</kernel.core.version>
 		<kernel.bom.version>1.3.1-rc.1</kernel.bom.version>
 		<java.version>21</java.version>
@@ -144,11 +143,6 @@
 			<groupId>io.mosip.preregistration</groupId>
 			<artifactId>pre-registration-core</artifactId>
 			<version>${pre.registration.core.version}</version>
-		</dependency>
-		<dependency>
-			<groupId>io.mosip.preregistration</groupId>
-			<artifactId>pre-registration-application-service</artifactId>
-			<version>${pre.registration.application.service.version}</version>
 		</dependency>
 		<dependency>
 			<groupId>io.mosip.kernel</groupId>
